+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Changelogs starts from version 0.1.3
 
 ## Released
 
+### [0.1.14]
+- register comments on fields in DBView models
+- fixed removing field comments in views
+
 ### [0.1.13]
 - Fix dependency issue (on psycopg) introduced in 0.1.12.